requirements.txt
python-telegram-bot==20.6
SQLAlchemy==2.0.22
alembic==1.12.0
python-dotenv==1.0.0
geopy==2.4.0
psycopg2-binary==2.9.9
loguru==0.7.2
APScheduler==3.10.4
pillow==10.0.1
pytest==7.4.2
pytest-asyncio==0.21.1 
watchdog
psutil